Blocked process count is now shown from perf counters and other tweaks
As we're heading towards real time data, the count of blocked processes is now shown from perf counters collected every 10 seconds rather then xes collected every 1 minute. Xes xml can be heavy to process so doing it every 10 seconds would not be a good choice. This means that the actual blocking chains will not be available to view immediately when the count goes up. There's going to be up to 1 minute delay.