Create test database on the flight rather than in a project
I made a full circle with this. We only need the test database to test blocking because SQLWATCH database has RCSI enabled which reduces blocking. I created the SQLWATCH-TESTER with the intention of moving test related stuff from SQLWATCH database to SQLWATCH-TESTER but I will move it all to Pester instead of T-SQL.